layer list refactor
