+++ elements/layercontrol/src/components/layerList.js
@@ -1,10 +1,10 @@
 import { LitElement, html } from "lit";
 import { when } from "lit/directives/when.js";
 import { repeat } from "lit/directives/repeat.js";
-import { checkProperties, createSortable, getLayerType } from "../helpers";
+import { getLayerType } from "../helpers";
 import "./layer";
 import "./layerGroup";
-import _debounce from "lodash.debounce";
+import { firstUpdatedMethod, updateMethod } from "../methods/layer-list";
 /**
  * Display of a list of layers
  *
@@ -70,34 +70,12 @@ export class EOxLayerControlLayerList extends LitElement {
     this.noShadow = true;
   }
 
-  #handleLayersChangeLength = () => {
-    this.#debHandleLayersChangeLength();
-  };
-  #debHandleLayersChangeLength = _debounce(() => {
-    this.requestUpdate();
-    this.dispatchEvent(new CustomEvent("changed", { bubbles: true }));
-  }, 50);
-
   firstUpdated() {
-    if (this.layers) {
-      checkProperties(this.layers, this.idProperty, this.titleProperty);
-      createSortable(
-        this.renderRoot.querySelector("ul"),
-        this.layers,
-        this.idProperty,
-        this
-      );
-    }
+    firstUpdatedMethod(this);
   }
 
   updated() {
-    if (!this.layers) {
-      return;
-    }
-    if (this.layers.hasListener("change:length")) {
-      this.layers?.un("change:length", this.#handleLayersChangeLength);
-    }
-    this.layers.on("change:length", this.#handleLayersChangeLength);
+    updateMethod(this);
   }
 
   createRenderRoot() {
@@ -105,6 +83,13 @@ export class EOxLayerControlLayerList extends LitElement {
   }
 
   render() {
+    const layers = this.layers
+      .getArray()
+      .filter(
+        (l) => !l.get("layerControlHide") && !l.get("layerControlOptional")
+      )
+      .reverse();
+
     return html`
       <style>
         ${this.#styleBasic}
@@ -115,13 +100,7 @@ export class EOxLayerControlLayerList extends LitElement {
           this.layers,
           () => html`
             ${repeat(
-              this.layers
-                .getArray()
-                .filter(
-                  (l) =>
-                    !l.get("layerControlHide") && !l.get("layerControlOptional")
-                )
-                .reverse(),
+              layers,
               (layer) => layer,
               (layer) => html`
                 <li
@@ -167,7 +146,6 @@ export class EOxLayerControlLayerList extends LitElement {
   }
 
   #styleBasic = ``;
-
   #styleEOX = `
     ul {
       padding: 0;
