+++ elements/layercontrol/src/methods/layer-list/index.js
@@ -1,0 +1,2 @@
+export { default as firstUpdatedMethod } from "./first-updated";
+export { default as updateMethod } from "./update";
